requirements.txt
slack-sdk>=3.26.0
slack-bolt>=1.18.0
python-dotenv>=1.0.0 